Added Serial Command, first alpha version :-)
Connect a USB-Serial interface to RxD1 and TxD1 and GND, 9600 bps.
Then type "TABCD" and hit Return to show "ABCD".
Type "C" and Return to clear the display.

You have to enter >= 4 characters after the "T" command, and only
the first 4 are shown in the first row.
The next version will allow more and less than 4 characters and also
x/y positioning.